refactor: Remove legacy top panel
