+++ operate/client/src/App/ProcessInstance/index.tsx
@@ -20,7 +20,7 @@ import {instanceHistoryModificationStore} from 'modules/stores/instanceHistoryMo
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 import {flowNodeTimeStampStore} from 'modules/stores/flowNodeTimeStamp';
 import {ProcessInstanceHeader} from './ProcessInstanceHeader';
-import {TopPanel} from './TopPanel/v2';
+import {TopPanel} from './TopPanel';
 import {BottomPanel, ModificationFooter, Buttons} from './styled';
 import {FlowNodeInstanceLog} from './FlowNodeInstanceLog/v2';
 import {Button, Modal} from '@carbon/react';
